refactor(ui): extract AIEngine helpers and error recovery
AIEngineContext.tsx: 557 → 416 lines.

  - contexts/ai/engineHelpers.ts: pure utilities (backendDisplayName,
    QUANT_LABELS, resolveBackendChain, quantFromModelName,
    buildReadyStatus)
  - contexts/ai/modelErrorRecovery.ts: the 85-line all-backends-failed
    toast logic (suggest an alternative quantization, offer fp32
    download, fall back to opening settings)

Public API surface of AIEngineContext is unchanged.
