+++ packages/ui/src/contexts/AIEngineContext.tsx
@@ -17,7 +17,6 @@ import {
   AnalysisQueue,
   pickConfig,
   probeEnvironment,
-  type AutoPick,
   type BackendId,
   type Engine,
   type Quantization,
@@ -32,6 +31,14 @@ import type { AISettings } from '../types/game';
 import { loadModelBuffer, modelIdFromName, resolveWasmPath } from './ai/engineLoader';
 import { tryEngineChain, type ChainStepEvent } from './ai/engineChain';
 import { idleStatus, isReady, type EngineStatus } from './ai/engineStatus';
+import {
+  QUANT_LABELS,
+  backendDisplayName,
+  buildReadyStatus,
+  quantFromModelName,
+  resolveBackendChain,
+} from './ai/engineHelpers';
+import { showModelErrorRecoveryToast } from './ai/modelErrorRecovery';
 
 // Module-level singletons survive StrictMode double-mount in dev. Without
 // these, mounting the provider twice would tear down a half-initialized
@@ -72,60 +79,6 @@ export function useAIEngine(): AIEngineContextValue {
 
 export function useAIEngineOptional(): AIEngineContextValue | null {
   return useContext(AIEngineContext);
-}
-
-/** User-friendly backend display names for toasts. */
-function backendDisplayName(backend: string): string {
-  switch (backend) {
-    case 'webgpu':
-    case 'webgpu-gc':
-      return 'GPU';
-    case 'native':
-    case 'native-cpu':
-      return 'Native';
-    case 'pytorch':
-      return 'PyTorch GPU';
-    case 'wasm':
-      return 'CPU';
-    case 'webnn':
-      return 'WebNN';
-    default:
-      return backend.toUpperCase();
-  }
-}
-
-const QUANT_LABELS: Record<ModelQuantization, string> = {
-  fp32: 'Full Quality',
-  fp16: 'Balanced',
-  uint8: 'Compact',
-};
-
-/**
- * Decide which backend chain to use:
- *  - settings.backend === 'auto' → full auto-pick (preferred default)
- *  - explicit setting → start from that backend, fall through the rest
- */
-function resolveBackendChain(settings: AISettings, autoPick: AutoPick): BackendId[] {
-  const explicit = settings.backend;
-  if (!explicit || explicit === 'auto') {
-    return autoPick.backendChain;
-  }
-  // Map old backend ids to BackendId; 'native' → 'native-gpu'
-  const mapped: BackendId =
-    explicit === 'native'
-      ? 'native-gpu'
-      : explicit === 'webnn' || explicit === 'webgl'
-        ? 'wasm'
-        : (explicit as BackendId);
-  // Start from explicit, then fall through the auto chain (de-duped).
-  return [mapped, ...autoPick.backendChain.filter(b => b !== mapped)];
-}
-
-/** Quantization label inferred from a model name (best effort). */
-function quantFromModelName(name: string): ModelQuantization {
-  if (/\.fp16\.|-fp16/i.test(name)) return 'fp16';
-  if (/\.uint8\.|-quant/i.test(name)) return 'uint8';
-  return 'fp32';
 }
 
 export const AIEngineProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
@@ -458,100 +411,6 @@ export const AIEngineProvider: React.FC<{ children: React.ReactNode }> = ({ chil
 
   return <AIEngineContext.Provider value={value}>{children}</AIEngineContext.Provider>;
 };
-
-function buildReadyStatus(): EngineStatus {
-  // We don't preserve auto-pick reasoning across module reloads; rebuild
-  // a minimal ready status. The provider re-derives full reasoning on
-  // its next initialize() call, which happens on first settings change.
-  return { phase: 'ready', backend: 'unknown', quantization: 'fp32', reasoning: '' };
-}
-
-/**
- * When all backends fail for a model, surface a recovery toast that
- * either suggests an alternative quantization (downloaded if possible,
- * else offers to download fp32) or offers to open the settings panel.
- */
-function showModelErrorRecoveryToast(
-  errorMessage: string,
-  selectedModelId: string | null,
-  modelLibrary: Array<{
-    id: string;
-    isDownloaded: boolean;
-    baseModelIndex?: number;
-    quantization?: ModelQuantization;
-  }>,
-  setSelectedModelId: (id: string | null) => void,
-  downloadModel: (id: string) => Promise<void>,
-  setAIConfigOpen: (open: boolean) => void,
-  showToast: (
-    message: string,
-    type: 'success' | 'error' | 'info',
-    action?: { label: string; onClick: () => void }
-  ) => void,
-  t: (key: string, opts?: Record<string, string>) => string
-): void {
-  const isFp16Error = /float16|fp16|Float16/.test(errorMessage);
-
-  if (selectedModelId) {
-    const parsed = parseModelId(selectedModelId);
-    if (parsed) {
-      const currentQuant = parsed.quantization;
-      const alternatives: ModelQuantization[] = (
-        ['fp32', 'uint8', 'fp16'] as ModelQuantization[]
-      ).filter(q => q !== currentQuant);
-
-      for (const alt of alternatives) {
-        const altId = getModelId(parsed.baseModelIndex, alt);
-        const altModel = modelLibrary.find(m => m.id === altId);
-        if (altModel?.isDownloaded) {
-          showToast(
-            isFp16Error
-              ? t('aiConfig.modelIncompatible', { quant: QUANT_LABELS[currentQuant] })
-              : t('aiConfig.allBackendsFailed'),
-            'error',
-            {
-              label: t('aiConfig.switchToModel', { quant: QUANT_LABELS[alt] }),
-              onClick: () => setSelectedModelId(altId),
-            }
-          );
-          return;
-        }
-      }
-
-      const fp32Id = getModelId(parsed.baseModelIndex, 'fp32');
-      const fp32Model = modelLibrary.find(m => m.id === fp32Id);
-      if (fp32Model && !fp32Model.isDownloaded) {
-        showToast(
-          isFp16Error
-            ? t('aiConfig.modelIncompatible', { quant: QUANT_LABELS[currentQuant] })
-            : t('aiConfig.allBackendsFailed'),
-          'error',
-          {
-            label: t('aiConfig.downloadAndSwitch', { quant: QUANT_LABELS['fp32'] }),
-            onClick: async () => {
-              try {
-                await downloadModel(fp32Id);
-                setSelectedModelId(fp32Id);
-              } catch {
-                showToast(t('aiConfig.modelDownloadFailed'), 'error');
-              }
-            },
-          }
-        );
-        return;
-      }
-    }
-  }
-
-  showToast(
-    isFp16Error ? t('aiConfig.modelIncompatibleGeneric') : t('aiConfig.allBackendsFailed'),
-    'error',
-    {
-      label: t('aiConfig.openSettings'),
-      onClick: () => setAIConfigOpen(true),
-    }
-  );
-}
 
 // Re-export the supporting types so they're discoverable from this module.
 export type { EngineStatus, BackendId, Quantization };
